move sais package under seg
